Favorites always play & Loading overlay
You only need to single click to play a favorite. Favorites will always
play right away (instead of only playing if it loaded all information
first). And when a favorite is selected a loading screen will appear and
last an extra 1.5 seconds. This is because the event "play-favorite"
doesn't actually end when the favorite is playing (and should be updated
to).